Authenticaton code implemented
Implemented authentication code so that user the must be in
ADMIN/CUSTOMER group to be able to get client information, and must
be in ADMIN group to be able to create a client.

Updated test cases to the new authentication code logic.

Updated Dockerfile setup to add 2 application users, john and root to
customer and admin group respectively.

Updated rest.yaml for the new 401 unathorized logic.

Updated jboss configuration.

Setup basic authentication.

Updated test case to deal with the new role logic

Updated test cases to deal with the new role logic for showing client
account information when the user is in role ADMIN.

Added authorization logic for adminCall and customerCall for
ADMIN and CUSTOMER role.