Added MySQL database container
Added MySQL database container instead of an embedded h2 database to
be used by the Wildfly for demonstration purpose.

Updated the persistence location for the test classes that uses the
EntityManager under the hood to continue depend on the h2 embedded
in memory database.

Updated the integration test to deal with the added mysql database
container.

Updated README.md about the concept to generate encrypted property
values with help of jasypt library; for this commit, the passwords
are in plain text for demonstration purpose.