Added encrypted datasource properties
Added 'encrypted' datasource properties to demonstrate how sensitive
property values can be retrieved with help of an encryption library
such as JASYPT.

Added JASYPT encrypted environment variables to be used to
for packaging the ear file with decrypted values on runtime.

Updated README.md file for what is required to install the application
in a container Docker environment. Installation documentation without
Docker is no longer supported, takes to much time to maintain.

Updates integration test to deal with the new datasource property
values.